moving searching collections to the home page
